+++ Reqnroll.Utils/IFileSystem.cs
@@ -7,6 +7,7 @@ namespace Reqnroll.Utils
     {
         bool FileExists(string path);
         bool DirectoryExists(string path);
+        void CreateDirectory(string path);
         string[] GetDirectories(string path);
         string[] GetDirectories(string path, string searchPattern);
         string[] GetDirectories(string path, string searchPattern, SearchOption searchOptions);