+++ Tests/Reqnroll.GeneratorTests/VirtualFileSystem.cs
@@ -214,5 +214,13 @@ namespace Reqnroll.GeneratorTests
                 throw new ArgumentException("Illegal characters in path", "searchPattern");
             }
         }
+
+        public void CreateDirectory(string path)
+        {
+            if (!DirectoryExists(path))
+            {
+                Directory.CreateDirectory(path);
+            }
+        }
     }
 }